Create Fascinating_number_or_not.java
Problem: Input an integer and check whether it is Fascinating number or not. Let's take any number (n) say 327 and check whether the given number is fascinating or not. On multiplying the given number (n) by 2 and 3, we get:
327*2=654
327*3=981
Now, concatenate the above results to the given number (n). 
"327"+"654"+"981" = "327654981"
We observe that the result contains all the digits from 1 to 9, exactly once, Hence, the given number 327 is a fascinating number.